requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
click==7.0                # via flask
flask==1.1.1
itsdangerous==1.1.0       # via flask
jinja2==2.10.3            # via flask
markupsafe==1.1.1         # via jinja2
prometheus-client==0.7.1
python-dotenv==0.10.3
werkzeug==0.16.0          # via flask
